Bruker proptypes fra rot i stedet for å ha det i de forksjellige pakkene
