+++ packages/fakta-arbeidsforhold/src/components/arbeidsforholdDetaljer/PersonAksjonspunktText.jsx
@@ -7,8 +7,8 @@ import { AksjonspunktHelpText, VerticalSpacer } from '@fpsak-frontend/shared-com
 import { DDMMYYYY_DATE_FORMAT, getKodeverknavnFn } from '@fpsak-frontend/utils';
 import kodeverkTyper from '@fpsak-frontend/kodeverk/src/kodeverkTyper';
 
+import { arbeidsforholdPropType } from '@fpsak-frontend/prop-types';
 import arbeidsforholdKilder from '../../kodeverk/arbeidsforholdKilder';
-import { arbeidsforholdPropType } from '../../propTypes/arbeidsforholdPropType';
 
 const utledPermisjonValues = (permisjon, getKodeverknavn) => {
   const kodeverknavn = getKodeverknavn(permisjon.type);
